+++ tools/ios.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import subprocess
+import common_compiler_flags
 from SCons.Variables import *
 
 if sys.version_info < (3,):
@@ -104,3 +105,5 @@ def generate(env):
     env.Append(LINKFLAGS=["-isysroot", env["IOS_SDK_PATH"], "-F" + env["IOS_SDK_PATH"]])
 
     env.Append(CPPDEFINES=["IOS_ENABLED", "UNIX_ENABLED"])
+
+    common_compiler_flags.generate(env)
